Move Parser::command
